feat: implement result pattern for user service, introduce user not found exception, and add update user DTO

+++ src/modules/user/services/user.service.ts
@@ -1,6 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { Prisma, User } from 'generated/prisma/client';
+import { User } from 'generated/prisma/client';
+import { Result } from 'src/core/result';
 import { IUserRepository } from 'src/domain/repositories/user.repository.interface';
+import { UpdateUserDto } from '../dto/update-user.dto';
+import { UserNotFoundException } from '../exceptions/user.exceptions';
 
 @Injectable()
 export class UserService {
@@ -8,19 +11,34 @@ export class UserService {
     @Inject('IUserRepository') private readonly userRepository: IUserRepository,
   ) {}
 
-  async findAll(): Promise<User[]> {
-    return this.userRepository.findAll();
+  async findAll(): Promise<Result<User[]>> {
+    const users = await this.userRepository.findAll();
+    return Result.ok(users);
   }
 
-  async findOne(id: string): Promise<User | null> {
-    return this.userRepository.findOne(id);
+  async findOne(id: string): Promise<Result<User>> {
+    const user = await this.userRepository.findOne(id);
+    if (!user) {
+      return Result.fail(new UserNotFoundException());
+    }
+    return Result.ok(user);
   }
 
-  async update(id: string, data: Prisma.UserUpdateInput): Promise<User> {
-    return this.userRepository.update(id, data);
+  async update(id: string, data: UpdateUserDto): Promise<Result<User>> {
+    const user = await this.userRepository.findOne(id);
+    if (!user) {
+      return Result.fail(new UserNotFoundException());
+    }
+    const updatedUser = await this.userRepository.update(id, data);
+    return Result.ok(updatedUser);
   }
 
-  async delete(id: string): Promise<User> {
-    return this.userRepository.delete(id);
+  async delete(id: string): Promise<Result<User>> {
+    const user = await this.userRepository.findOne(id);
+    if (!user) {
+      return Result.fail(new UserNotFoundException());
+    }
+    const deletedUser = await this.userRepository.delete(id);
+    return Result.ok(deletedUser);
   }
 }
